lambda/requirements.txt
boto3-stubs[ecs]>=1.35.97
pydantic>=2.10.5
